+++ src/main/java/dao/ConsumerDao.java
@@ -6,8 +6,8 @@ import entiy.Consumer;
 public interface ConsumerDao {
 
 	Consumer quaryConsumerPassword(@Param("telNum") String telNum, @Param("password") String password);
-	Consumer quaryConsumerInfo(@Param("consumerId") long consumerId);
-	int insertConsumerInfo(@Param("consumerId") long conmsumerId);
-	int changeConsumerInfo(@Param("consumerId") long consumerId);
+	Consumer quaryConsumerInfo(@Param("consumerId") int consumerId);
+	int insertConsumerInfo(@Param("consumerId") int conmsumerId);
+	int changeConsumerInfo(@Param("consumerId") int consumerId);
 	
 }
